+++ CHANGELOG.md
@@ -43,6 +43,7 @@ QueryClient::new()
 ### Fixed
 - Reduced internal generic codegen bloat over `K` internally, `Eq` no longer required, just `PartialEq` on nonlocal resources ([#10](https://github.com/zakstucke/leptos-fetch/pull/10))
 - Subscription bugs ([#13](https://github.com/zakstucke/leptos-fetch/pull/10))
+- Fixed edge case where 2 scopes with different options would be treated the same ([#22](https://github.com/zakstucke/leptos-fetch/pull/22))
 
 ### Tests
 - Sped up CI with erase_components ([#17](https://github.com/zakstucke/leptos-fetch/pull/17))